[DEPYLER-0313] Add explicit type annotation for ambiguous numeric operations
Added explicit i32 cast before .abs() to avoid ambiguous numeric type
errors in range step calculations.

Problem:
- Expressions like `(-1).abs()` have ambiguous type {integer}
- Rust cannot infer if literal is i8, i16, i32, i64, or i128
- Range generation with negative steps failed to compile

Solution:
- Changed `(#step).abs()` to `(#step as i32).abs()`
- Explicit i32 type annotation resolves ambiguity
- Consistent with Python int → i32 mapping

Testing:
- All 453 core tests pass (zero regressions)
- Test case verified type annotation works
- No more "ambiguous numeric type" errors
- Clippy clean with -D warnings

Impact:
- Matrix 07_algorithms: ~6 errors → ~5 errors (estimated)
- Fixed 1 ambiguous type error
- Progress: 11/16 total errors fixed (68.75%)

Files Modified:
- crates/depyler-core/src/rust_gen/expr_gen.rs (line 1017)
  * Changed to `(#step as i32).abs()`
  * Added DEPYLER-0313 comment

Design Pattern:
- Explicit casting before method calls
- No ambiguity in numeric literal operations
- Type annotation at point of use
- i32 is standard for Python int mapping

+++ crates/depyler-core/src/rust_gen/expr_gen.rs
@@ -1013,7 +1013,8 @@ impl<'a, 'b> ExpressionConverter<'a, 'b> {
         // Python: range(10, 0, -1) → Rust: (0..10).rev()
         Ok(parse_quote! {
             {
-                let step = (#step).abs() as usize;
+                // DEPYLER-0313: Cast to i32 before abs() to avoid ambiguous numeric type
+                let step = (#step as i32).abs() as usize;
                 if step == 0 {
                     panic!("range() arg 3 must not be zero");
                 }
